refactor: 배틀페이지(index)에서 battle:attacked/defensed 이벤트 수신 처리 및 관련 이펙트 처리 로직을  useBattleTimeline훅으로 분리
- 새로운 useBattleTimeline 훅 생성 및 결과 이펙트 표시 로직 이 훅으로 이동
